luxe/requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
fpdf2
